Fixed Web Serial API Integration
The EKG or web serial api (I am not sure which) sends the data as a utf8 char. So you have to store each char into a string and once you get a new line that is one data point. From there you need to parse that string into a float.